LUN-2334 Added option to disable cache clears during long operations.

+++ menus/menu_pool.py
@@ -83,10 +83,11 @@ class MenuPool(object):
         if all:
             cache_keys = CacheKey.objects.get_keys()
         else:
-            cache_keys = CacheKey.objects.get_keys(site_id, language)        
+            cache_keys = CacheKey.objects.get_keys(site_id, language)
         to_be_deleted = cache_keys.distinct().values_list('key', flat=True)
-        cache.delete_many(to_be_deleted)
-        cache_keys.delete()
+        if to_be_deleted:
+            cache.delete_many(to_be_deleted)
+            cache_keys.delete()
     
     def register_menu(self, menu):
         from menus.base import Menu
